Updated server exit
Now exiting isn't done on any random key press. This was to fix an issue
relating to systemd starting the service. I never got to the bottom of it, but
I think systemd was sending something via stdin to the process. Instead of
figuring out what and why systemd was doing that. I decided it'd be easier just
to accept a smaller range of inputs.

Also (because atomic commits are for organised people):
 * Moved NUM_LINK_REDIRECTS to better place
 * Now use 302 redirect code (temporary is better than permanent, especially
   when testing or if the links change)
 * Improved the examples to include "http" and "https", because they are
   required